Move conferences to own directory
- remove from generic folder
- create new urls file
