+++ meetings/api/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
-from .models import Submission, Conference
+from .models import Submission
 
 admin.site.register(Submission)
-admin.site.register(Conference)
