+++ meetings/api/models.py
@@ -14,26 +14,8 @@ class Submission(models.Model):
     title = models.CharField(max_length=100)
     contributors = models.ManyToManyField(User, blank=True)
     description = models.TextField()
-    conference = models.ForeignKey('conference')
+    conference = models.ForeignKey('conferences.Conference')
     approved = models.NullBooleanField(blank=True)
 
     class Meta:
         ordering = ('date_created',)
-
-class Conference(models.Model):
-    created = models.DateTimeField(auto_now_add=True)
-    modified = models.DateTimeField(auto_now=True)
-    title = models.CharField(max_length=100)
-    site_url = models.URLField(blank=True)
-    city = models.CharField(max_length=100)
-    state = models.CharField(max_length=100)
-    country = CountryField(blank_label='(select country)')
-    event_start = models.DateTimeField(blank=True, null=True)
-    event_end = models.DateTimeField(blank=True, null=True)
-    submission_start = models.DateTimeField(blank=True, null=True)
-    submission_end = models.DateTimeField(blank=True, null=True)
-    logo_url = models.URLField(blank=True)
-    description = models.TextField(blank=True, max_length=500)
-
-    class Meta:
-        ordering = ('created',)
