+++ meetings/api/serializers.py
@@ -1,7 +1,8 @@
 from django.contrib.auth.models import User, Group
-from api.models import Submission, Conference
+from api.models import Submission
 from rest_framework import serializers as ser
-from django_countries.fields import CountryField
+
+from conferences.models import Conference
 
 class UserSerializer(ser.HyperlinkedModelSerializer):
     class Meta:
@@ -20,24 +21,6 @@ class AuthenticationSerializer(ser.Serializer):
     def validate(self, data):
         return data
 
-class ConferenceSerializer(ser.ModelSerializer):
-    title = ser.CharField(required=True)
-    city = ser.CharField()
-    state = ser.CharField()
-    country = CountryField() #get country_dict working later
-    start = ser.DateTimeField(source='event_start', required=False)
-    end = ser.DateTimeField(source='event_end', required=False)
-    submissionStart = ser.DateTimeField(source='submission_start', required=False)
-    submissionEnd = ser.DateTimeField(source='submission_end', required=False)
-    logoUrl = ser.URLField(source='logo_url', allow_blank=True)
-    description = ser.CharField(allow_blank=True)
-    siteUrl = ser.URLField(source='site_url')
-
-    # Later on add tags and sponsors back
-    class Meta:
-        model = Conference
-        fields = ('created', 'modified', 'id', 'title', 'siteUrl', 'city',
-                'state', 'country', 'start', 'end', 'submissionStart', 'submissionEnd', 'logoUrl', 'description')
 
 class SubmissionSerializer(ser.HyperlinkedModelSerializer):
     conference = ser.PrimaryKeyRelatedField(queryset=Conference.objects.all())
