+++ meetings/conferences/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from conferences.models import Conference
+
+admin.site.register(Conference)
