+++ meetings/meetings/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'corsheaders',
     'api',
+    'conferences',
 ]
 
 REST_FRAMEWORK = {
