+++ meetings/meetings/urls.py
@@ -14,8 +14,7 @@ router.register(r'users', views.UserViewSet)
 urlpatterns = [
     url(r'^', include(router.urls)),
     url(r'^users/(?P<pk>[0-9]+)/$', views.UserDetail, name='user-detail'),
-    url(r'^conferences/$', views.ConferenceList.as_view(), name='Conference-list'),
-    url(r'^conferences/(?P<pk>[0-9]+)/$', views.ConferenceDetail.as_view(), name='conference_detail'),
+    url(r'^conferences/', include('conferences.urls', namespace='conferences')),
     url(r'^conferences/(?P<conference_id>[0-9]+)/submissions/$', views.SubmissionList.as_view(), name='submission-list'),
     url(r'^conferences/(?P<conference_id>[0-9]+)/submissions/(?P<submission_id>[0-9]+)/$',
         views.SubmissionDetail.as_view(), name='submission-detail'),
